requirements.txt
Flask==3.1.2
# https://bootstrap-flask.readthedocs.io/en/stable/
bootstrap-flask==2.5.0
# https://flask-sqlalchemy.readthedocs.io/en/stable/
Flask-SQLAlchemy==3.1.1
# https://flask-migrate.readthedocs.io/en/latest/
Flask-Migrate==4.1.0
# https://flask-login.readthedocs.io/en/latest/
flask-login==0.6.3
# https://flask-wtf.readthedocs.io/en/1.2.x/
Flask-WTF==1.2.2
# Para permitir validação de emails no WTForms
# https://github.com/JoshData/python-email-validator
email-validator==2.3.0
# Para gerar JWTs
# https://pyjwt.readthedocs.io/en/stable/
PyJWT==2.10.1
# Para enviar email
# https://postmarker.readthedocs.io/en/latest/
postmarker==1.0
